+++ word-wrap/history.md
@@ -1,7 +1,87 @@
 # Word Wrap Kata
-[First Attempt](#first-attempt)
+* [Last](#last-time)
+* [Second](#second-time)
+* [First](#first-time)
 
-## Second Attempt
+## Last Time
+<!-- [Commit ...](https://github.com/median-man/javascript-katas/commit/) -->
+
+This kata went quickly. I did the third practice on the following day. It took less than an hour from start to finish. I did not look up any code. I nearly let a bug get by me. I didn't add a test case for `columns <= 0` until the end of the kata.
+
+### Code
+```javasciprt
+function wordWrap(text, columns = 0) {
+  if (!text || columns <= 0) return '';
+  if (text.length <= columns) return text;
+
+  let wrapAt = text.lastIndexOf(' ', columns);
+  if (wrapAt === -1) wrapAt = columns;
+
+  const firstLine = text.substr(0, wrapAt);
+  const remaining = wordWrap(text.substr(wrapAt).trim(), columns);
+  return `${firstLine}\n${remaining}`;
+}
+
+module.exports = wordWrap;
+```
+
+### Tests
+```javasciprt
+/* eslint prefer-arrow-callback: 0, func-names: 0 */
+const { assert } = require('chai');
+const wordWrap = require('./word-wrap.js');
+
+describe('wordWrap', function () {
+  it('is a function', function () {
+    assert.isFunction(wordWrap);
+  });
+
+  describe('degenerate cases', function () {
+    it('returns an empty string when passed text is falsey', function () {
+      assert.equal(wordWrap(), '', 'when text undefined');
+      assert.equal(wordWrap(null, 5), '', 'when text is null');
+    });
+
+    it('returns the text when the text is shorter than columns', function () {
+      assert.equal(wordWrap('dune', 7), 'dune');
+    });
+
+    it('returns an empty string when columns is falsey', function () {
+      assert.equal(wordWrap('spice'), '');
+    });
+
+    it('returns an empty string when columns is less than 0', function () {
+      assert.equal(wordWrap('spice', -1), '');
+    });
+  });
+
+  describe('when passed a single word', function () {
+    it('wraps the word when column is less than the length of the word', function () {
+      assert.equal(wordWrap('fremen', 5), 'freme\nn');
+    });
+
+    it('wraps the word multiple times', function () {
+      assert.equal(wordWrap('arrakis', 3), 'arr\naki\ns');
+    });
+  });
+
+  describe('when passed two or more words', function () {
+    it('wraps at the word boundry', function () {
+      assert.equal(wordWrap('paul dune', 5), 'paul\ndune');
+    });
+
+    it('wraps at first word when it fits on a line', function () {
+      assert.equal(wordWrap('paul atreides', 10), 'paul\natreides');
+    });
+
+    it('wraps after multiple words when they fit on a line', function () {
+      assert.equal(wordWrap('i must not fear', 8), 'i must\nnot fear');
+    });
+  });
+});
+```
+
+## Second Time
 [Commit 4f04521be4ab67a8367a5f1a3c8a42c10d141502](https://github.com/median-man/javascript-katas/commit/4f04521be4ab67a8367a5f1a3c8a42c10d141502)
 
 My second attempt at solving the Word Wrap Kata went far, far faster than the first. After the first, I read a blog post on the word wrap algorithm and then immediately began working on the kata from scratch. I still got a little stuck transitioning to line breaks one subsequent spaces. But I worked through it relatively quickly. I'm surprised at how different the code I wrote looks even after refactoring. Although, the algorithm still has the same basic recursion.
@@ -71,7 +151,7 @@ describe('wordWrap', function () {
 });
 ```
 
-## First Attempt
+## First Time
 [Commit 9900688af7b336efc1c6f7cab20a171d59325ef1](https://github.com/median-man/javascript-katas/commit/9900688af7b336efc1c6f7cab20a171d59325ef1)
 
 My first attempt at solving the Word Wrap Kata went okay. I wanted to do it cold without any Google cheating aside from a few lookups for Javascript String and Array methods as wells as getting reacquainted with using and configuring the Mocha/Chai test suite and assertion libraries.